requirements.txt
pillow==11.0.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5-stubs==5.15.6.0
PyQt5_sip==12.15.0
python-barcode==0.15.1
